Reorganize files, fix semantics and CSS.
