+++ src/components/grid-aware/Footer/Footer.module.css
@@ -1,11 +1,7 @@
 /* Background bleed */
 
 .bleedWrapper {
-  display: grid;
-  grid-template-columns: 1fr;
-  grid-template-rows: auto 0;
   margin: auto;
-  max-width: var(--grid-bleed-max-width);
 }
 
 /* Main grid */
@@ -13,7 +9,6 @@
 .gridParent {
   column-gap: var(--grid-column-gap);
   display: grid;
-  grid-template-rows: 120px 380px 78px;
 }
 
 /* Override grid item default of auto so that item contents cannot stretch the grid itself */
@@ -29,229 +24,6 @@
   .gridParent {
     column-gap: var(--grid-column-gap);
     display: grid;
-    grid-template-rows: 190px auto auto;
   }
 }
 
-/* Top Row */
-
-.gridTopRow {
-  align-content: center;
-  background-color: var(--color-black);
-  display: flex;
-  grid-column: 1 / span 12;
-  grid-row: 1;
-  justify-content: center;
-}
-
-.subscribeContainer {
-  align-items: baseline;
-  align-self: center;
-  display: flex;
-  justify-content: center;
-}
-
-.subscribeMessage {
-  color: var(--color-white);
-  font: var(--font-subtitle);
-  padding-right: 20px;
-  width: 340px;
-}
-
-.gridTopRow a {
-  padding: 14px 40px;
-}
-
-.gridTopRow a:hover {
-  top: 0;
-}
-
-.formInput {
-  display: flex;
-  flex-direction: row;
-}
-
-.formInput input {
-  border: 0;
-  color: var(--color-black);
-  font: var(--font-body-small);
-  padding: 16px 0 15px 17px;
-  width: 291px;
-}
-
-@media (--tablet-and-down) {
-  .gridTopRow {
-    padding-bottom: 43px;
-    padding-top: 44px;
-  }
-
-  .subscribeMessage {
-    padding-bottom: 22px;
-  }
-
-  .subscribeContainer {
-    display: unset;
-    flex-direction: column;
-  }
-
-  .subscribeMessage {
-    padding-right: 0;
-  }
-
-  .formInput {
-    display: unset;
-  }
-
-  .formInput input {
-    width: 155px;
-  }
-}
-
-/* Middle Row */
-
-.gridMiddleRow {
-  background-color: var(--color-deep-blue);
-  display: flex;
-  flex-direction: row;
-  grid-column: 1 / span 12;
-  grid-row: 2;
-  justify-content: space-between;
-  padding: 60px 75px 0;
-}
-
-.gridMiddleLeftContainer {
-  text-align: left;
-}
-
-.gridMiddleRightContainer {
-  text-align: right;
-}
-
-.footerLinksContainer {
-  display: flex;
-  flex-direction: row;
-}
-
-.footerLinksColumn {
-  display: flex;
-  flex-direction: column;
-  text-align: left;
-}
-
-.footerLinksColumn:not(:last-child) {
-  margin-right: 70px;
-}
-
-.footerLinks {
-  color: var(--color-white);
-  display: flex;
-  flex-direction: column;
-  font: var(--font-body-small);
-  margin-top: 30px;
-}
-
-.footerLinks a {
-  color: inherit;
-  text-decoration: none;
-}
-
-.footerLinks a:not(:first-child) {
-  margin-top: 15px;
-}
-
-.footerLinksTitle {
-  color: var(--color-white);
-  font: var(--font-caption);
-}
-
-.sealsContainer {
-  margin-top: 60px;
-}
-
-.sealsContainer img:not(:last-child) {
-  margin-right: 39px;
-}
-
-.socialMediaLinks {
-  margin-top: 40px;
-}
-
-.socialMediaLinks a:not(:last-child) {
-  margin-right: 20px;
-}
-
-@media (--tablet-and-down) {
-  .gridMiddleRow {
-    flex-direction: column;
-    justify-content: center;
-    padding: 40px 0 60px;
-  }
-
-  .footerLinksContainer {
-    flex-direction: column;
-    text-align: center;
-  }
-
-  .gridMiddleLeftContainer {
-    text-align: center;
-  }
-
-  .gridMiddleRightContainer {
-    margin-top: 60px;
-  }
-
-  .footerLinksColumn:not(:last-child) {
-    margin-right: 0;
-    text-align: center;
-  }
-
-  .footerLinksColumn:not(:first-child) {
-    margin-right: 0;
-    margin-top: 40px;
-    text-align: center;
-  }
-
-  .footerLinksColumn {
-    text-align: center;
-  }
-
-  .sealsContainer {
-    text-align: center;
-  }
-}
-
-/* Bottom Row */
-
-.gridBottomRow {
-  background-color: var(--color-midnight-blue);
-  color: var(--color-gray-300);
-  display: grid;
-  font-family: var(--font-family-lato);
-  font-size: 14px;
-  font-weight: 300;
-  grid-column: 1 / span 12;
-  grid-row: 3;
-  line-height: 16.8px;
-  padding: 0 60px;
-}
-
-.organizationInfo {
-  align-items: center;
-  display: flex;
-  flex-direction: row;
-  justify-content: space-between;
-}
-
-@media (--tablet-and-down) {
-  .gridBottomRow {
-    padding: 30px 0 35px;
-  }
-
-  .organizationInfo {
-    flex-direction: column;
-  }
-
-  .organizationInfo div:nth-child(2) {
-    margin-top: 15px;
-  }
-}
