+++ src/components/grid-aware/Footer/SubscriptionBlock/SubscriptionBlock.module.css
@@ -1,0 +1,80 @@
+/* Background bleed */
+
+.bleedWrapper {
+  display: grid;
+  grid-template-columns: 1fr minmax(var(--grid-content-min-width), var(--grid-content-max-width)) 1fr;
+  margin: auto;
+}
+
+/* Top Row */
+
+.subscriptionBlock {
+  background-color: var(--color-black);
+  display: flex;
+  grid-column: 1 / span 12;
+  justify-content: center;
+  padding: 35px;
+}
+
+.subscribeContainer {
+  align-items: baseline;
+  display: flex;
+}
+
+.subscribeMessage {
+  color: var(--color-white);
+  font: var(--font-subtitle);
+  margin-right: 20px;
+  width: 340px;
+}
+
+.subscriptionBlock a {
+  padding: 14px 40px;
+}
+
+.subscriptionBlock a:hover {
+  top: 0;
+}
+
+.formInput {
+  display: flex;
+  flex-direction: row;
+}
+
+.formInput input {
+  border-color: #e0e0e0;
+  color: var(--color-black);
+  font: var(--font-body-small);
+  padding: 16px 16px 15px 17px;
+  width: 291px;
+}
+
+@media (--tablet-and-down) {
+  .bleedWrapper {
+    display: block;
+    grid-template-columns: 1fr;
+    margin: 0;
+  }
+
+  .subscribeMessage {
+    margin-bottom: 22px;
+  }
+
+  .subscribeContainer {
+    display: flex;
+    flex-direction: column;
+    align-items: normal;
+  }
+
+  .subscribeMessage {
+    margin-right: 0;
+  }
+
+  .formInput {
+    display: unset;
+  }
+
+  .formInput input {
+    width: 155px;
+  }
+}
